@@ -104,6 +104,7 @@ class WindowManager {
     await _channel.invokeMethod('ensureInitialized');
   }
 
+  /// You can call this to remove the window frame (title bar, outline border, etc), which is basically everything except the Flutter view, also can call setTitleBarStyle(TitleBarStyle.normal) or setTitleBarStyle(TitleBarStyle.hidden) to restore it.
   Future<void> setAsFrameless() async {
     await _channel.invokeMethod('setAsFrameless');
   }
@@ -114,6 +115,13 @@ class WindowManager {
     VoidCallback? callback,
   ]) async {
     await _channel.invokeMethod('waitUntilReadyToShow');
+
+    if (options?.titleBarStyle != null) {
+      await setTitleBarStyle(
+        options!.titleBarStyle!,
+        windowButtonVisibility: options.windowButtonVisibility ?? true,
+      );
+    }
 
     if (await isFullScreen()) await setFullScreen(false);
     if (await isMaximized()) await unmaximize();
@@ -138,12 +146,6 @@ class WindowManager {
       await setSkipTaskbar(options!.skipTaskbar!);
     }
     if (options?.title != null) await setTitle(options!.title!);
-    if (options?.titleBarStyle != null) {
-      await setTitleBarStyle(
-        options!.titleBarStyle!,
-        windowButtonVisibility: options.windowButtonVisibility ?? true,
-      );
-    }
 
     if (callback != null) {
       callback();
@@ -570,7 +572,7 @@ class WindowManager {
     bool windowButtonVisibility = true,
   }) async {
     final Map<String, dynamic> arguments = {
-      'titleBarStyle': describeEnum(titleBarStyle),
+      'titleBarStyle': titleBarStyle.name,
       'windowButtonVisibility': windowButtonVisibility,
     };
     await _channel.invokeMethod('setTitleBarStyle', arguments);
@@ -696,7 +698,7 @@ class WindowManager {
   /// Sets the brightness of the window.
   Future<void> setBrightness(Brightness brightness) async {
     final Map<String, dynamic> arguments = {
-      'brightness': describeEnum(brightness),
+      'brightness': brightness.name,
     };
     await _channel.invokeMethod('setBrightness', arguments);
   }
@@ -733,7 +735,7 @@ class WindowManager {
     await _channel.invokeMethod<bool>(
       'startResizing',
       {
-        'resizeEdge': describeEnum(resizeEdge),
+        'resizeEdge': resizeEdge.name,
         'top': resizeEdge == ResizeEdge.top ||
             resizeEdge == ResizeEdge.topLeft ||
             resizeEdge == ResizeEdge.topRight,
