@@ -28,7 +28,26 @@
 #define STATE_FULLSCREEN_ENTERED 3
 #define STATE_DOCKED 4
 
-#define DWMWA_USE_IMMERSIVE_DARK_MODE 19
+/// Window attribute that enables dark mode window decorations.
+///
+/// Redefined in case the developer's machine has a Windows SDK older than
+/// version 10.0.22000.0.
+/// See:
+/// https://docs.microsoft.com/windows/win32/api/dwmapi/ne-dwmapi-dwmwindowattribute
+#ifndef DWMWA_USE_IMMERSIVE_DARK_MODE
+#define DWMWA_USE_IMMERSIVE_DARK_MODE 20
+#endif
+
+constexpr const wchar_t kWindowClassName[] = L"FLUTTER_RUNNER_WIN32_WINDOW";
+
+/// Registry key for app theme preference.
+///
+/// A value of 0 indicates apps should use dark mode. A non-zero or missing
+/// value indicates apps should use light mode.
+constexpr const wchar_t kGetPreferredBrightnessRegKey[] =
+    L"Software\\Microsoft\\Windows\\CurrentVersion\\Themes\\Personalize";
+constexpr const wchar_t kGetPreferredBrightnessRegValue[] =
+    L"AppsUseLightTheme";
 
 #define APPBAR_CALLBACK WM_USER + 0x01;
 
@@ -251,7 +270,7 @@ void WindowManager::Blur() {
 }
 
 bool WindowManager::IsFocused() {
-  return GetMainWindow() == GetActiveWindow();
+  return GetMainWindow() == GetForegroundWindow();
 }
 
 void WindowManager::Show() {
@@ -323,6 +342,10 @@ bool WindowManager::IsMinimized() {
 }
 
 void WindowManager::Minimize() {
+  if (IsFullScreen()) {  // Like chromium, we don't want to minimize fullscreen
+                         // windows
+    return;
+  }
   HWND mainWindow = GetMainWindow();
   WINDOWPLACEMENT windowPlacement;
   GetWindowPlacement(mainWindow, &windowPlacement);
@@ -352,7 +375,7 @@ int WindowManager::IsDocked() {
 
 double WindowManager::GetDpiForHwnd(HWND hWnd) {
   auto monitor = MonitorFromWindow(hWnd, MONITOR_DEFAULTTONEAREST);
-  UINT newDpiX = 96; // Default values
+  UINT newDpiX = 96;  // Default values
   UINT newDpiY = 96;
 
   // Dynamically load shcore.dll and get the GetDpiForMonitor function address
@@ -362,12 +385,13 @@ double WindowManager::GetDpiForHwnd(HWND hWnd) {
     typedef HRESULT (*GetDpiForMonitor)(HMONITOR, int, UINT*, UINT*);
 
     GetDpiForMonitor GetDpiForMonitorFunc =
-      (GetDpiForMonitor)GetProcAddress(shcore, "GetDpiForMonitor");
+        (GetDpiForMonitor)GetProcAddress(shcore, "GetDpiForMonitor");
 
     if (GetDpiForMonitorFunc) {
       // Use the loaded function if available
       const int MDT_EFFECTIVE_DPI = 0;
-      if (FAILED(GetDpiForMonitorFunc(monitor, MDT_EFFECTIVE_DPI, &newDpiX, &newDpiY))) {
+      if (FAILED(GetDpiForMonitorFunc(monitor, MDT_EFFECTIVE_DPI, &newDpiX,
+                                      &newDpiY))) {
         // If it fails, set the default values again
         newDpiX = 96;
         newDpiY = 96;
@@ -375,8 +399,8 @@ double WindowManager::GetDpiForHwnd(HWND hWnd) {
     }
     FreeLibrary(shcore);
   }
-  return ((double) newDpiX);
-} 
+  return ((double)newDpiX);
+}
 
 void WindowManager::Dock(const flutter::EncodableMap& args) {
   HWND mainWindow = GetMainWindow();
@@ -468,7 +492,6 @@ void PASCAL WindowManager::AppBarQuerySetPos(HWND hwnd,
 }
 
 BOOL WindowManager::RegisterAccessBar(HWND hwnd, BOOL fRegister) {
-
   APPBARDATA abd;
 
   // Specify the structure size and handle to the appbar.
@@ -550,8 +573,8 @@ void WindowManager::SetFullScreen(const flutter::EncodableMap& args) {
 
   // Previously inspired by how Chromium does this
   // https://src.chromium.org/viewvc/chrome/trunk/src/ui/views/win/fullscreen_handler.cc?revision=247204&view=markup
-  // Instead, we use a modified implementation of how the media_kit package implements this
-  // (we got permission from the author, I believe)
+  // Instead, we use a modified implementation of how the media_kit package
+  // implements this (we got permission from the author, I believe)
   // https://github.com/alexmercerind/media_kit/blob/1226bcff36eab27cb17d60c33e9c15ca489c1f06/media_kit_video/windows/utils.cc
 
   // Save current window state if not already fullscreen.
@@ -563,26 +586,31 @@ void WindowManager::SetFullScreen(const flutter::EncodableMap& args) {
     g_title_bar_style_before_fullscreen = title_bar_style_;
   }
 
-  if (isFullScreen) {
+  g_is_window_fullscreen = isFullScreen;
+
+  if (isFullScreen) {  // Set to fullscreen
     ::SendMessage(mainWindow, WM_SYSCOMMAND, SC_MAXIMIZE, 0);
     if (!is_frameless_) {
-    auto monitor = MONITORINFO{};
-    auto placement = WINDOWPLACEMENT{};
-    monitor.cbSize = sizeof(MONITORINFO);
-    placement.length = sizeof(WINDOWPLACEMENT);
-    ::GetWindowPlacement(mainWindow, &placement);
-    ::GetMonitorInfo(::MonitorFromWindow(mainWindow, MONITOR_DEFAULTTONEAREST),
-                     &monitor);
-    ::SetWindowLongPtr(mainWindow, GWL_STYLE, g_style_before_fullscreen & ~WS_OVERLAPPEDWINDOW);
-    ::SetWindowPos(mainWindow, HWND_TOP, monitor.rcMonitor.left,
-                   monitor.rcMonitor.top, monitor.rcMonitor.right,
-                   monitor.rcMonitor.bottom,
-                   SWP_NOOWNERZORDER | SWP_FRAMECHANGED);
+      auto monitor = MONITORINFO{};
+      auto placement = WINDOWPLACEMENT{};
+      monitor.cbSize = sizeof(MONITORINFO);
+      placement.length = sizeof(WINDOWPLACEMENT);
+      ::GetWindowPlacement(mainWindow, &placement);
+      ::GetMonitorInfo(
+          ::MonitorFromWindow(mainWindow, MONITOR_DEFAULTTONEAREST), &monitor);
+      ::SetWindowLongPtr(mainWindow, GWL_STYLE,
+                         g_style_before_fullscreen & ~WS_OVERLAPPEDWINDOW);
+      ::SetWindowPos(mainWindow, HWND_TOP, monitor.rcMonitor.left,
+                     monitor.rcMonitor.top,
+                     monitor.rcMonitor.right - monitor.rcMonitor.left,
+                     monitor.rcMonitor.bottom - monitor.rcMonitor.top,
+                     SWP_NOOWNERZORDER | SWP_FRAMECHANGED);
     }
-  } else {
+  } else {  // Restore from fullscreen
     if (!g_maximized_before_fullscreen)
       Restore();
-    ::SetWindowLongPtr(mainWindow, GWL_STYLE, g_style_before_fullscreen | WS_OVERLAPPEDWINDOW);
+    ::SetWindowLongPtr(mainWindow, GWL_STYLE,
+                       g_style_before_fullscreen | WS_OVERLAPPEDWINDOW);
     if (::IsZoomed(mainWindow)) {
       // Refresh the parent mainWindow.
       ::SetWindowPos(mainWindow, nullptr, 0, 0, 0, 0,
@@ -590,8 +618,8 @@ void WindowManager::SetFullScreen(const flutter::EncodableMap& args) {
                          SWP_FRAMECHANGED);
       auto rect = RECT{};
       ::GetClientRect(mainWindow, &rect);
-      auto flutter_view =
-          ::FindWindowEx(mainWindow, nullptr, kFlutterViewWindowClassName, nullptr);
+      auto flutter_view = ::FindWindowEx(mainWindow, nullptr,
+                                         kFlutterViewWindowClassName, nullptr);
       ::SetWindowPos(flutter_view, nullptr, rect.left, rect.top,
                      rect.right - rect.left, rect.bottom - rect.top,
                      SWP_NOACTIVATE | SWP_NOZORDER);
@@ -606,8 +634,6 @@ void WindowManager::SetFullScreen(const flutter::EncodableMap& args) {
           SWP_NOACTIVATE | SWP_NOZORDER);
     }
   }
-
-  g_is_window_fullscreen = isFullScreen;
 }
 
 void WindowManager::SetAspectRatio(const flutter::EncodableMap& args) {
@@ -838,22 +864,16 @@ void WindowManager::SetAlwaysOnTop(const flutter::EncodableMap& args) {
 }
 
 bool WindowManager::IsAlwaysOnBottom() {
-    return is_always_on_bottom_;
+  return is_always_on_bottom_;
 }
 
 void WindowManager::SetAlwaysOnBottom(const flutter::EncodableMap& args) {
   is_always_on_bottom_ =
       std::get<bool>(args.at(flutter::EncodableValue("isAlwaysOnBottom")));
 
-  SetWindowPos(
-    GetMainWindow(),
-    is_always_on_bottom_ ? HWND_BOTTOM : HWND_NOTOPMOST,
-    0,
-    0,
-    0,
-    0,
-    SWP_NOMOVE | SWP_NOSIZE
-  );
+  SetWindowPos(GetMainWindow(),
+               is_always_on_bottom_ ? HWND_BOTTOM : HWND_NOTOPMOST, 0, 0, 0, 0,
+               SWP_NOMOVE | SWP_NOSIZE);
 }
 
 std::string WindowManager::GetTitle() {
@@ -1000,14 +1020,21 @@ void WindowManager::SetOpacity(const flutter::EncodableMap& args) {
 }
 
 void WindowManager::SetBrightness(const flutter::EncodableMap& args) {
-  std::string brightness =
-      std::get<std::string>(args.at(flutter::EncodableValue("brightness")));
+  DWORD light_mode;
+  DWORD light_mode_size = sizeof(light_mode);
+  LSTATUS result =
+      RegGetValue(HKEY_CURRENT_USER, kGetPreferredBrightnessRegKey,
+                  kGetPreferredBrightnessRegValue, RRF_RT_REG_DWORD, nullptr,
+                  &light_mode, &light_mode_size);
 
-  const BOOL is_dark_mode = brightness == "dark";
-
-  HWND hWnd = GetMainWindow();
-  DwmSetWindowAttribute(hWnd, DWMWA_USE_IMMERSIVE_DARK_MODE, &is_dark_mode,
-                        sizeof(is_dark_mode));
+  if (result == ERROR_SUCCESS) {
+    std::string brightness =
+        std::get<std::string>(args.at(flutter::EncodableValue("brightness")));
+    HWND hWnd = GetMainWindow();
+    BOOL enable_dark_mode = light_mode == 0 && brightness == "dark";
+    DwmSetWindowAttribute(hWnd, DWMWA_USE_IMMERSIVE_DARK_MODE,
+                          &enable_dark_mode, sizeof(enable_dark_mode));
+  }
 }
 
 void WindowManager::SetIgnoreMouseEvents(const flutter::EncodableMap& args) {
